% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAnalysis.R
\name{clust_ana}
\alias{clust_ana}
\title{Cluster analysis}
\usage{
clust_ana(ans, k, distm = "euclidean", hclustm = "ward.D",
  na.rm = FALSE)
}
\arguments{
\item{ans}{data matrix}

\item{k}{number of clusters}

\item{distm}{Dissimilarity index, partial match to "\code{manhattan}", "\code{euclidean}"(default), "\code{canberra}", "\code{bray}", "\code{kulczynski}", "\code{jaccard}", "\code{gower}", "\code{altGower}", "\code{morisita}", "\code{horn}", "\code{mountford}", "\code{raup}" , "\code{binomial}", "\code{chao}", "\code{cao}" or "\code{mahalanobis}".}

\item{hclustm}{the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "\code{ward.D}"(default), "\code{ward.D2}", "\code{single}", "\code{complete}", "\code{average}" (= UPGMA), "\code{mcquitty}" (= WPGMA), "\code{median}" (= WPGMC) or "\code{centroid}" (= UPGMC).}

\item{na.rm}{Pairwise deletion of missing observations when computing dissimilarities.}
}
\description{
Does cluster analysis and report an index of cluster
}
\examples{
\dontrun{
data("SegData")
library(dplyr)
library(vegan)
cind <- clust_ana(dat, k=4)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
