% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-glasso.R
\name{cv_glasso}
\alias{cv_glasso}
\title{Cross-validation for group lasso logistic regression}
\usage{
cv_glasso(trainx, trainy, nlam = 100, type = "link", kfold = 10,
  na_action = na.pass)
}
\arguments{
\item{trainx}{a data frame where samples are in rows and features are in columns}

\item{trainy}{a numeric or factor vector containing the outcome for each sample}

\item{nlam}{number of lambda values. The default is 100}

\item{type}{the type of prediction. \code{type = 'link'} is on the scale of linear predictors (default), whereas \code{type = 'response'} is on the scale of the response  variable, i.e. \code{type = 'response'} applies the inverse link function to the linear predictors.}

\item{kfold}{number of folds - default is 10. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. The default is 10.}

\item{na_action}{function determining what should be done with missing values when predicting new data during cross validation. The default is to predict NA.}
}
\value{
an object of class "\code{cv_glasso}" is returned, which is a list with the ingredients of the cross-validation fit.
}
\description{
Does k-fold cross-validation for group lasso logistic regression and returns a list object
}
\examples{
\dontrun{
data("sim1_da1")
trainx = dplyr::select(sim1_da1, -y)
trainy = sim1_da1$y
# index of the group
index <- gsub("\\\\..*", "", names(trainx))
# nlam is the number of values of tuning variable
nlam <- 10
# type of prediction
type = "link"
# number of cross-validation folds
kfold <- 10
cv_fit <- cv_glasso(trainx, trainy, nlam = nlam, kfold = kfold)
str(cv_fit)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
