% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-dat.R
\name{impute_dat}
\alias{impute_dat}
\title{Impute missing values}
\usage{
impute_dat(dat, method)
}
\arguments{
\item{dat}{a data frame.}

\item{method}{imputation method. Need to be one of these: \code{zero}, \code{mean}, \code{Inf0}, \code{code99}}
}
\description{
Impute missing values
}
\examples{
\dontrun{
library(DataScienceR)
data("SegData")
impdat <- SegData[, c("income", "house")]
# summary(SegData)
impdat = impute_dat(impdat, method = "mean")
summary(impdat)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
