% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeterRevAct.R
\name{MeterRevAct}
\alias{MeterRevAct}
\title{Calculate metered revenue accounting matrix}
\usage{
MeterRevAct(datall_revenue, type, dates)
}
\arguments{
\item{datall_revenue}{subscription data, need to have columns:  user_id, Product, Effective_Start, one_time_charge}

\item{type}{product type}

\item{dates}{a date vector (have to be date type)}
}
\value{
A revenue accounting matrix is returned for each date.
}
\description{
It calculates the metered accounting matrix given date and subscription data. It only returns one row for each date, i.e. it will sum up all the products.
}
\examples{
\dontrun{
dates <- seq(as.Date("2018-02-01"), lubridate::ceiling_date(Sys.Date(), unit = "month"), "months") - days(1)
dates
metered_growth_matrix = MeterRevAct(datall_revenue = meter_revenue, type=one_time_charge_product, dates = dates)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
