% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeterRevActUserProduct.R
\name{MeterRevActUserProduct}
\alias{MeterRevActUserProduct}
\title{Calculate metered revenue accounting matrix per day, user and product}
\usage{
MeterRevActUserProduct(datall_revenue, dates, type)
}
\arguments{
\item{datall_revenue}{subscription data, need to have columns:  user_id, Product, Effective_Start, one_time_charge}

\item{dates}{a date vector (have to be date type)}

\item{type}{a vector of product names}
}
\value{
A metered revenue accounting matrix is returned for each day, user and product
}
\description{
It calculates the accounting matrix for each user and product given date, subscription data and product category.
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
