% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrrAtDateType.R
\name{mrrAtDateType}
\alias{mrrAtDateType}
\title{Calculate MRR for a specific date and product type}
\usage{
mrrAtDateType(time_point, subs, type = NULL)
}
\arguments{
\item{time_point}{a date value (have to be date type)}

\item{subs}{subscription data, need to have columns: account_id,user_id,Product,Effective_Start,Effective_End,MRR}

\item{type}{product type, if "NULL", it will return MRR for each product}
}
\value{
a vector of MRR is returned
}
\description{
It calculates the MRR given date, subscription date, and product type
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
