% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_ks_test.R
\name{pairwise_ks_test}
\alias{pairwise_ks_test}
\title{Pairwise Kolmogorov-Smirnov Test}
\usage{
pairwise_ks_test(value, group, n_min = 50, warning = 0,
  alternative = "two.sided")
}
\arguments{
\item{value}{a numeric vector of data values}

\item{group}{a group indicator}

\item{n_min}{The minimum number of observations in each group. Group(s) with observation less than \code{n_min} will be removed. "\code{n_min=50}"(default)}

\item{warning}{sets the handling of warning messages. If \code{warning} is negative all warnings are ignored. If \code{warning} is zero (the default) warnings are stored until the top–level function returns. If 10 or fewer warnings were signalled they will be printed otherwise a message saying how many were signalled. An object called last.warning is created and can be printed through the function warnings. If \code{warning} is one, warnings are printed as they occur. If warn is two or larger all warnings are turned into errors.}

\item{alternative}{indicates the alternative hypothesis and must be one of "\code{two.sided}" (default), "\code{less}", or "\code{greater}". You can specify just the initial letter of the value, but the argument name must be give in full. See ‘Details’ for the meanings of the possible values.}
}
\value{
Pairwise Kolmogorov-Smirnov Test p-value Matrix
}
\description{
Perform Pairwise Multiple Comparisons Using Kolmogorov-Smirnov Test
}
\details{
Missing values are silently omitted from x and (in the two-sample case) y.

The possible values "\code{two.sided}", "\code{less}" and "\code{greater}" of alternative specify the null hypothesis that the true distribution function of x is equal to, not less than or not greater than the hypothesized distribution function (one-sample case) or the distribution function of y (two-sample case), respectively. This is a comparison of cumulative distribution functions, and the test statistic is the maximum difference in value, with the statistic in the "greater" alternative being \eqn{D^+ = max[F_x(u) - F_y(u)]}. Thus in the two-sample case alternative = "greater" includes distributions for which x is stochastically smaller than y (the CDF of x lies above and hence to the left of that for y), in contrast to \code{t.test} or \code{wilcox.test}.
}
\examples{
\dontrun{
data("iris")
value<-iris$Sepal.Length
group<-iris$Species
pairwise_ks_test(value,group,warning = -1)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
