% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv_glasso.R
\name{plot.cv_glasso}
\alias{plot.cv_glasso}
\title{plot the cross-validation curve produced by  "\code{cv_glasso}"}
\usage{
\method{plot}{cv_glasso}(x, type.measure = "auc", ...)
}
\arguments{
\item{x}{fitted "\code{cv_glasso}" object}

\item{type.measure}{criteria to use for cross-validation. Currently three options. The default is \code{type.measure = "auc"} which gives area under the ROC curve. \code{type.measure = "loglike"} computes the log-likelihood score in Meier et al2008. \code{type.measure = "maxco"} computes the maximum correlation coefficient in Yeo and Burge.}
}
\value{
A plot is produced, and nothing is returned.
}
\description{
Plots the cross-validation curve as a function of the lambda values used.
}
\examples{
\dontrun{
data("sim1_da1")
trainx = dplyr::select(sim1_da1, -y)
trainy = sim1_da1$y
# index of the group
index <- gsub("\\\\..*", "", names(trainx))
# nlam is the number of values of tuning variable
nlam <- 10
# type of prediction
type = "link"
# number of cross-validation folds
kfold <- 10
cv.fit <- cv_glasso(trainx, trainy, nlam = nlam, kfold = kfold)
plot.cv_glasso(cv.fit)
}

}
\references{
L. Meier, S. van de Geer, and P. Buhlmann, The group lasso for logistic regression, J. R. Stat. Soc. Ser. B Stat. Methodol. 70 (2008), pp. 53-71.

G.W. Yeo and C.B. Burge, Maximum entropy modeling of short sequence motifs with applications to RNA splicing signals, J. Computnl Biol. 11 (2004), pp. 475-494.
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
