% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rocTest.R
\name{plot.rocTest}
\alias{plot.rocTest}
\title{Plot ROC curve}
\usage{
\method{plot}{rocTest}(x, auto.legend = TRUE, ...)
}
\arguments{
\item{x}{rocTest object}
}
\description{
Plot ROC curve based on rocTest object
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
