% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tune_cutoff.R
\name{plot.tune.cutoff}
\alias{plot.tune.cutoff}
\title{Plot cutoff tuning process from "\code{tune_cutoff}"}
\usage{
\method{plot}{tune.cutoff}(x, pch = 20, cex = 1.5, col = "red", ...)
}
\arguments{
\item{x}{"\code{tune_cutoff}" object}

\item{pch}{type of cutoff points, default is 20}

\item{cex}{size of cutoff points, default is 1.5, col = "red"}

\item{col}{color of cutoff points, default is \code{col = "red"}}

\item{...}{other parameters in "\code{plot()}" function}
}
\description{
Plots cutoff tuning process
}
\examples{
\dontrun{
data("sim1_da1")
trainx = sim1_da1[,1:50]
trainy = sim1_da1$y
library(glmnet)
fit <- cv.glmnet(as.matrix(trainx), trainy, family = "binomial")
test <- predict(fit, as.matrix(trainx), type = "link", s = "lambda.min")
test <- as.vector(test)
summary(test)
likelihood <- c(0.2, 0.5, 0.8)
y <- trainy
x <- tune_cutoff(test = test, y = y, likelihood = likelihood)
str(x)
plot.tune.cutoff(x)
}

}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
