% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fitglasso.R
\name{predict_glasso}
\alias{predict_glasso}
\title{Predict method for grplasso objects}
\usage{
predict_glasso(object, newdata, type = c("link", "response"),
  na.action = na.pass, ...)
}
\arguments{
\item{object}{a \code{grplasso} object}

\item{newdata}{data.frame or design matrix of new observations}

\item{type}{the type of prediction. \code{type = "link"} is on the scale of linear predictors, whereas \code{type = "response"} is on the scale of the response variable, i.e. \code{type = "response"} applies the inverse link function to the linear predictors.
na.action    function determining what should be done with missing values in newdata. The default is to predict NA.}

\item{...}{other options to be passed to the predict function.}
}
\description{
Obtains predictions from a \code{grplasso} object.
}
