% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevAct_Q.R
\name{RevAct_Q}
\alias{RevAct_Q}
\title{Calculate revenue accounting matrix for each quarter}
\usage{
RevAct_Q(datall_revenue, type, dates)
}
\arguments{
\item{datall_revenue}{subscription data, need to have columns:  user_id, Product, Effective_Start, Effective_End, MRR}

\item{type}{product type}

\item{dates}{a date vector (have to be date type)}
}
\value{
A revenue accounting matrix is returned for each date.
}
\description{
It calculates the accounting matrix given date, subscription data, and product type. It only returns one row for each quarter, i.e. it will sum up all the products.
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
