% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevActUser.R
\name{RevActUser}
\alias{RevActUser}
\title{Calculate revenue accounting matrix for each user}
\usage{
RevActUser(datall_revenue, dates)
}
\arguments{
\item{datall_revenue}{subscription data, need to have columns:  user_id, Effective_Start, Effective_End, MRR}

\item{dates}{a date vector (have to be date type)}
}
\value{
A revenue accounting matrix is returned for each date and user
}
\description{
It calculates the accounting matrix given date, subscription data.
}
\examples{
\dontrun{
res = RevActUser(total_revenue, dates)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
