% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevActUserProduct.R
\name{RevActUserProduct}
\alias{RevActUserProduct}
\title{Calculate revenue accounting matrix per day, user and product}
\usage{
RevActUserProduct(datall_revenue, dates, dates_full, type)
}
\arguments{
\item{datall_revenue}{daily revenue data, need to have columns:  user_id, Product, active_date, MRR}

\item{dates}{a date vector (have to be date type)}

\item{dates_full}{a full range of dates}

\item{type}{a vector of product names}
}
\value{
A revenue accounting matrix is returned for each day, user and product
}
\description{
It calculates the accounting matrix for each user and product given date, subscription data and product category.
}
\examples{
\dontrun{
team <- c("netlify_team_premium_monthly","Teams","netlify_team_plus_monthly","netlify_team_plus_yearly")
res_0 = RevActUserProduct(total_revenue, dates, team)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
