% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevCntAct.R
\name{RevCntAct}
\alias{RevCntAct}
\title{Calculate revenue accounting matrix by counts}
\usage{
RevCntAct(datall_revenue, type, dates)
}
\arguments{
\item{datall_revenue}{subscription data, need to have columns:  user_id, Product, Effective_Start, Effective_End, MRR}

\item{type}{product type}

\item{dates}{a date vector (have to be date type)}
}
\value{
A revenue accounting matrix by count is returned for each date.
}
\description{
It calculates the accounting matrix by counts given date and subscription data. It only returns one row for each date, i.e. it will sum up all the products.
}
\examples{
\dontrun{
dates <- seq(as.Date("2018-02-01"), lubridate::ceiling_date(Sys.Date(), unit = "month"), "months") - days(1)
dates
growth_matrix_identity = RevCntAct(datall_revenue = total_revenue, type = c("Identity"), dates = dates)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
