% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocTest.R
\name{rocTest}
\alias{rocTest}
\title{Compute the area under the ROC curve and compare different AUC's}
\usage{
rocTest(y, x, L = NULL)
}
\arguments{
\item{y}{response}

\item{x}{prediction}

\item{L}{list to assign the comparisons}
}
\description{
This function compares the AUC of two correlated (or paired) or uncorrelated (un- paired) ROC curves.
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
