% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep-data.R
\name{sep_data}
\alias{sep_data}
\title{Separate data frame according to variable type}
\usage{
sep_data(dat)
}
\arguments{
\item{dat}{a data frame.}
}
\description{
Separate data frame according to variable type
}
