% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclass_eff_est.R
\name{subclass_eff_est}
\alias{subclass_eff_est}
\title{Estimation ATT and ATE after stratification}
\usage{
subclass_eff_est(object, y)
}
\arguments{
\item{object}{The output object from matchit. This is a required input.}

\item{y}{outcome variable. This is a required input.}
}
\value{
Fuction returns a list with the ATT and ATE estimates
}
\description{
After stratification, estimation the weighted mean difference. Return estimates of ATT and ATE
}
\examples{
\dontrun{
subclass_eff_est(m2.out, match.data(m2.out)$CORN_UNITS_CHG)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
