% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune-cutoff.R
\name{tune_cutoff}
\alias{tune_cutoff}
\title{Tune the cutoffs for different likelihood groups}
\usage{
tune_cutoff(test, y, likelihood)
}
\arguments{
\item{test}{a vector of predicted score from model}

\item{y}{a vector of response, \code{y} need to be the same length of \code{test}}

\item{likelihood}{a vector of likelihood values}
}
\value{
an object of class "\code{tune_cutoff}" is returned, which is a list with the ingredients of the tuning process.
}
\description{
Tune the cutoffs for different likelihood groups given predicted score, response and a vector of likelihood values
}
\examples{
\dontrun{
data("sim1_da1")
trainx = sim1_da1[,1:50]
trainy = sim1_da1$y
library(glmnet)
fit <- cv.glmnet(as.matrix(trainx), trainy, family = "binomial")
test <- predict(fit, as.matrix(trainx), type = "link", s = "lambda.min")
test <- as.vector(test)
summary(test)
likelihood <- c(0.2, 0.5, 0.8)
y <- trainy
x <- tune_cutoff(test = test, y = y, likelihood = likelihood)
str(x)
plot.tune.cutoff(x)
}

}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
