% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrrAtDateType.R
\name{mrrAtDateType}
\alias{mrrAtDateType}
\title{Calculate MRR for a specific date and product type}
\usage{
mrrAtDateType(time_point, subs, type)
}
\arguments{
\item{time_point}{a date value (have to be date type)}

\item{subs}{subscription data}

\item{type}{product type}
}
\value{
a number of MRR is returned
}
\description{
It calculates the MRR given date, subscription date, and product type
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
