% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevAct.R
\name{RevAct}
\alias{RevAct}
\title{Calculate revenue accounting matrix}
\usage{
RevAct(datall_revenue, type, dates)
}
\arguments{
\item{datall_revenue}{daily revenue data, need to have columns:  user_id, Product, active_date, MRR}

\item{type}{product type}

\item{dates}{a date vector (have to be date type)}
}
\value{
A revenue accounting matrix is returned for each date.
}
\description{
It calculates the accounting matrix given date and daily revenue data. It only returns one row for each date, i.e. it will sum up all the products.
}
\examples{
\dontrun{
dates <- seq(as.Date("2018-02-01"), lubridate::ceiling_date(Sys.Date(), unit = "month"), "months") - days(1)
dates
# need to get daily revenue from subscription data if you don't have that already
base_monthly = do.call("rbind", lapply(dates, function(time_point)
growth_matrix_identity = RevAct(datall_revenue = base_monthly, type = c("Identity"), dates = dates)
}
}
\author{
Hui Lin, \email{longqiman@gmail.com}
}
